+++ metrics/time.py
@@ -50,6 +50,44 @@ class Timer:
 
         last_section.end_section()
 
+    def to_dict(self):
+        json_dict                 = self.__dict__
+        json_dict_sections        = {}
+        json_dict_looped_sections = {}
+
+        for section, section_obj in self.sections.items():
+            json_dict_sections[section] = section_obj.to_dict()
+
+        for section, section_list in self.looped_sections.items():
+            json_dict_looped_sections[section] = [
+                s.to_dict() for s in section_list
+            ]
+
+        json_dict["sections"] = json_dict_sections
+        json_dict["looped_sections"] = json_dict_looped_sections
+        json_dict["function_time"] = self.function_time.to_dict()
+
+        return json_dict
+
+    @classmethod
+    def create_from_dict(cls, raw_dict):
+        timer = cls(
+            raw_dict["function_name"], raw_dict["call_id"]
+        )
+        timer.function_time = TimerSection.create_from_dict(raw_dict["function_time"])
+        timer.sections = {
+            section: TimerSection.create_from_dict(section_dict) 
+            for section, section_dict in raw_dict["sections"].items()
+        }
+        timer.looped_sections = {
+            section: [
+                TimerSection.create_from_dict(section_dict)
+                for section_dict in section_list
+            ] 
+            for section, section_list in raw_dict["looped_sections"].items()
+        }
+        return timer
+
 
 class TimerSection:
     """
@@ -67,15 +105,23 @@ class TimerSection:
             raise TimerException("Attempted to get elapsed of a non-finished section.")
         return self.end - self.start
 
+    def to_dict(self): return self.__dict__
+
+    @classmethod
+    def create_from_dict(cls, raw_dict):
+        section = cls()
+        section.start = raw_dict["start"]
+        section.end   = raw_dict["end"]
+        return section
+
 
 class TimerException(Exception): pass
 
 
 def uses_timer(func):
 
-    timer = Timer(func.__name__, str(uuid.uuid4()))
-
     def wrapper(*args, **kwargs):
+        timer = Timer(func.__name__, str(uuid.uuid4()))
         return func(timer, *args, **kwargs)
 
     return wrapper